API Reference

Index

InferOpt.AbstractPerturbed

InferOpt.AbstractPerturbed

InferOpt.FenchelYoungLoss

InferOpt.Interpolation

InferOpt.IsBaseLoss

InferOpt.IsRegularized

InferOpt.PerturbedAdditive

InferOpt.PerturbedComposition

InferOpt.PerturbedMultiplicative

InferOpt.SPOPlusLoss

InferOpt.StructuredSVMLoss

InferOpt.ZeroOneBaseLoss

InferOpt.compute_maximizer

InferOpt.compute_y_and_F

InferOpt.half_square_norm

InferOpt.isproba

InferOpt.isprobadist

InferOpt.one_hot_argmax

InferOpt.positive_part

InferOpt.ranking

InferOpt.shannon_entropy

InferOpt.simplex_projection_and_support

InferOpt.soft_argmax

InferOpt.sparse_argmax

Interpolation

Reference

Differentiation of Blackbox Combinatorial Solvers

InferOpt.Interpolation — Type

Interpolation{F}

Piecewise-linear interpolation of a black-box optimizer.

Fields

maximizer::F: underlying argmax function

λ::Float64: smoothing parameter (smaller = more faithful approximation, larger = more informative gradients)

Reference: https://arxiv.org/abs/1912.02175

Smart "Predict, then Optimize"

Reference

Smart "Predict, then Optimize"

InferOpt.SPOPlusLoss — Type

SPOPlusLoss{F}

Convex surrogate of the Smart "Predict-then-Optimize" loss.

Fields

maximizer::F: linear maximizer function of the form θ ⟼ ŷ(θ) = argmax θᵀy

α::Float64: convexification parameter

Reference: https://arxiv.org/abs/1710.08005

Structured Support Vector Machines

Reference

Structured learning and prediction in computer vision, Chapter 6

InferOpt.IsBaseLoss — Type

IsBaseLoss{L}

Trait-based interface for loss functions δ(y, y_true), which are the base of the more complex StructuredSVMLoss.

For δ::L to comply with this interface, the following methods must exist:

(δ)(y, y_true)

compute_maximizer(δ, θ, α, y_true)

Available implementations

ZeroOneBaseLoss

InferOpt.compute_maximizer — Method

compute_maximizer(δ, θ, α, y_true)

Compute argmax_y {δ(y, y_true) + α θᵀ(y - y_true)} to deduce the gradient of a StructuredSVMLoss.

InferOpt.StructuredSVMLoss — Type

StructuredSVMLoss{L}

Loss associated with the Structured Support Vector Machine.

ℓ(θ, y_true) = max_y {δ(y, y_true) + α θᵀ(y - y_true)}

Fields

base_loss::L: of the IsBaseLoss trait

α::Float64

Reference: http://www.nowozin.net/sebastian/papers/nowozin2011structured-tutorial.pdf (Chapter 6)

InferOpt.ZeroOneBaseLoss — Type

ZeroOneBaseLoss

0-1 loss for multiclass classification: δ(y, y_true) = 0 if y = y_true, and 1 otherwise.

Regularized optimizers

Reference

Learning with Fenchel-Young Losses

InferOpt.IsRegularized — Type

IsRegularized{P}

Trait-based interface for regularized prediction functions ŷ(θ) = argmax {θᵀy - Ω(y)}.

For predictor::P to comply with this interface, the following methods must exist:

(predictor)(θ)

compute_regularization(predictor, y)

Available implementations

one_hot_argmax

soft_argmax

sparse_argmax

InferOpt.soft_argmax — Method

soft_argmax(z)

Soft argmax activation function s(z) = (e^zᵢ / ∑ e^zⱼ)ᵢ.

Corresponds to regularized prediction on the probability simplex with entropic penalty.

InferOpt.simplex_projection_and_support — Method

simplex_projection_and_support(z)

Compute the Euclidean projection p of z on the probability simplex (also called sparse_argmax), and the indicators s of its support.

Reference: https://arxiv.org/abs/1602.02068.

InferOpt.sparse_argmax — Method

sparse_argmax(z)

Compute the Euclidean projection of the vector z onto the probability simplex.

Corresponds to regularized prediction on the probability simplex with square norm penalty.

InferOpt.half_square_norm — Method

half_square_norm(x)

Compute the squared Euclidean norm of x and divide it by 2.

InferOpt.isproba — Method

isproba(x)

Check whether x ∈ [0,1].

InferOpt.isprobadist — Method

isprobadist(p)

Check whether the elements of p are nonnegative and sum to 1.

InferOpt.one_hot_argmax — Method

one_hot_argmax(z)

One-hot encoding of the argmax function.

InferOpt.positive_part — Method

positive_part(x)

Compute max(x,0).

InferOpt.ranking — Method

ranking(θ[; rev])

Compute the vector r such that rᵢ is the rank of θᵢ in θ.

InferOpt.shannon_entropy — Method

shannon_entropy(p)

Compute the Shannon entropy of a probability distribution: H(p) = -∑ pᵢlog(pᵢ).

Perturbed optimizers

Reference

Learning with Differentiable Perturbed Optimizers

InferOpt.AbstractPerturbed — Type

AbstractPerturbed{F}

Differentiable perturbation of a black-box optimizer.

Available subtypes

PerturbedAdditive{F}

PerturbedMultiplicative{F}

Required fields

maximizer::F: underlying argmax function

ε::Float64: noise scaling parameter

rng::AbstractRNG: random number generator

seed::Union{Nothing,Int}: random seed

nb_samples::Int: number of random samples for Monte-Carlo computations

Required methods

(perturbed)(θ, Z; kwargs...)

compute_y_and_F(perturbed, θ, Z; kwargs...)

Optional methods

rrule(perturbed, θ; kwargs...)

InferOpt.AbstractPerturbed — Method

(perturbed)(θ, Z; kwargs...)

InferOpt.compute_y_and_F — Method

compute_y_and_F(perturbed, θ, Z; kwargs...)

InferOpt.PerturbedAdditive — Type

PerturbedAdditive{F}

Differentiable normal perturbation of a black-box optimizer: the input undergoes θ -> θ + εZ where Z ∼ N(0, I).

See also: AbstractPerturbed{F}.

Reference: https://arxiv.org/abs/2002.08676

InferOpt.PerturbedComposition — Type

PerturbedComposition{F,P<:AbstractPerturbed{F},G}

Composition of a differentiable perturbed black-box optimizer with an arbitrary function.

Suitable for direct regret minimization (learning by experience) when said function is a cost.

Fields

perturbed::P: underlying AbstractPerturbed{F} wrapper

g::G: function taking an array y and some kwargs as inputs

The method rrule(perturbed_composition, θ; kwargs...) must be implemented individually for each specific type P.

Base.:∘ — Method

∘(g, perturbed)

Create a PerturbedComposition object from perturbed and g.

ChainRulesCore.rrule — Method

rrule(perturbed_composition, θ; kwargs...)

InferOpt.PerturbedMultiplicative — Type

PerturbedMultiplicative{F}

Differentiable log-normal perturbation of a black-box optimizer: the input undergoes θ -> θ ⊙ exp[εZ - ε²/2] where Z ∼ N(0, I).

See also: AbstractPerturbed{F}.

Fenchel-Young losses

Reference

Learning with Fenchel-Young Losses

InferOpt.FenchelYoungLoss — Type

FenchelYoungLoss{P}

Fenchel-Young loss associated with a given regularized prediction function.

Fields

predictor::P: prediction function of the form ŷ(θ) = argmax {θᵀy - Ω(y)}

Reference: https://arxiv.org/abs/1901.02324

Implicit differentiation

Reference

Efficient and Modular Implicit Differentiation

Stay tuned!

This will soon be implemented thanks to the recent package ImplicitDifferentiation.jl.