API Reference

Index

InferOpt.FenchelYoungLoss

InferOpt.GeneralStructuredLoss

InferOpt.Interpolation

InferOpt.IsRegularizedPrediction

InferOpt.IsStructuredLossFunction

InferOpt.Perturbed

InferOpt.PerturbedCost

InferOpt.SPOPlusLoss

InferOpt.StructuredSVMLoss

InferOpt.ZeroOneLoss

InferOpt.half_square_norm

InferOpt.isproba

InferOpt.isprobadist

InferOpt.one_hot_argmax

InferOpt.positive_part

InferOpt.shannon_entropy

InferOpt.simplex_projection_and_support

InferOpt.softmax

InferOpt.sparsemax

Docstrings

InferOpt.FenchelYoungLoss — Type

FenchelYoungLoss{P}

Fenchel-Young loss associated with a given regularized prediction function.

Fields

predictor::P: prediction function, usually of the form θ ⟼ ŷ(θ) = argmax ⟨θ,y⟩ - Ω(y)

InferOpt.GeneralStructuredLoss — Type

GeneralStructuredLoss{F1, F2}

delta_loss(y, y_true)
maximizer(θ, α, y_true) = argmax_y (delta_loss(y, y_true) + α ⟨θ,  y - y_true⟩)

InferOpt.Interpolation — Type

Interpolation{F}

Piecewise-linear interpolation of a black-box optimizer.

Fields

maximizer::F: underlying argmax function

λ::Float64: smoothing parameter (smaller = more faithful approximation, larger = more informative gradients)

InferOpt.IsRegularizedPrediction — Type

IsRegularizedPrediction{P}

Trait-based interface for regularized prediction functions of the form θ ⟼ ŷ(θ) = argmax ⟨θ,y⟩ - Ω(y).

For a type P to comply with this interface, the following methods must exist:

(prediction::P)(θ)

compute_regularization(prediction::P, y)

We provide some special cases with explicit formulae, such as:

one_hot_argmax

softmax

sparsemax

InferOpt.IsStructuredLossFunction — Type

IsStructuredLossFunction{L}

Trait-based interface for structured loss functions of the form `(y, ytrue) ⟼ l(y, ytrue). You need a StructuredLossFunction in order build StructuredSVM losses.

For a type L to comply with this interface, the following methods must exist:

(loss::L)(y, y_true)

compute_maximizer(loss::L, θ, α, y_true)

it should return the following : argmaxy (loss(y, ytrue) + α ⟨θ, y - y_true⟩)

useful to compute the gradient of an associated SSVM loss.

We provide some special cases with explicit formulae, such as:

ZeroOneLoss

We also provide a generic wrapper GeneralStructuredLoss to build your own StructuredLossFunction.

InferOpt.Perturbed — Type

Perturbed{F}

Differentiable perturbation of a black-box optimizer.

Fields

maximizer::F: underlying argmax function

ε::Float64: noise scaling parameter

M::Int: number of noise samples for Monte-Carlo computations

InferOpt.PerturbedCost — Type

PerturbedCost{F,C}

Composition of a differentiable perturbed black-box optimizer with an arbitrary cost function.

Designed for direct regret minimization (learning by experience).

Fields

maximizer::F: underlying argmax function

ε::Float64: noise scaling parameter

M::Int: number of noise samples for Monte-Carlo computations

cost::C: a real-valued function taking a vector y and an instance as inputs

See also

Perturbed

InferOpt.SPOPlusLoss — Type

SPOPlusLoss{F}

Convex surrogate of the SPO loss.

Fields

maximizer::F: linear maximizer function of the form θ ⟼ ŷ(θ) = argmax ⟨θ,y⟩

α::Float64: convexification parameter

InferOpt.StructuredSVMLoss — Type

StructuredSVMLoss{L, R<:Real}

L should satisfy the IsStructuredLossFunction trait.

SSVMloss(θ, y_true) = max_y (l(y, y_true) + α [⟨θ, y⟩ - ⟨θ, y_true⟩])

InferOpt.ZeroOneLoss — Type

ZeroOneLoss

Zero-one loss for multiclass classification (ZeroOneLoss(y, y_true) equals 0 if y == y_true, else 1).

InferOpt.half_square_norm — Method

half_square_norm(x)

Compute the squared Euclidean norm of x and divide it by 2.

InferOpt.isproba — Method

isproba(x)

Check whether x ∈ [0,1].

InferOpt.isprobadist — Method

isprobadist(p)

Check whether the elements of p are nonnegative and sum to 1.

InferOpt.one_hot_argmax — Method

one_hot_argmax(θ)

One-hot encoding of the argmax function.

Corresponds to regularized prediction on the probability simplex with zero penalty.

InferOpt.positive_part — Method

positive_part(x)

Compute max(x,0).

InferOpt.shannon_entropy — Method

shannon_entropy(p)

Compute the Shannon entropy of a probability distribution: H(p) = -∑ pᵢlog(pᵢ).

InferOpt.simplex_projection_and_support — Method

simplex_projection_and_support(z)

Compute the Euclidean projection p of z on the probability simplex (also called sparsemax), and the indicators s of its support.

See https://arxiv.org/abs/1602.02068.

InferOpt.softmax — Method

softmax(θ)

Softmax function s(θ) = (e^θᵢ / ∑ e^θⱼ)ᵢ.

Corresponds to regularized prediction on the probability simplex with entropic penalty.

InferOpt.sparsemax — Method

sparsemax(z)

Project the vector z onto the probability simplex Δ in time O(d log d).

Implementation and chain rule from https://arxiv.org/abs/1602.02068.

Corresponds to regularized prediction on the probability simplex with square norm penalty.