API Reference

Probability distributions

InferOpt.FixedAtomsProbabilityDistribution — Type

FixedAtomsProbabilityDistribution{A,W}

Encodes a probability distribution with finite support and fixed atoms.

See compute_expectation to understand the name of this struct.

Fields

atoms::Vector{A}: elements of the support

weights::Vector{W}: probability values for each atom (must sum to 1)

InferOpt.FixedAtomsProbabilityDistribution — Method

FixedAtomsProbabilityDistribution(s::FrankWolfe.ActiveSet)

Construct a distribution from the active set generated by a Frank-Wolfe algorithm.

Base.rand — Method

rand([rng,] probadist)

Sample from the atoms of probadist according to their weights.

InferOpt.apply_on_atoms — Method

apply_on_atoms(post_processing, probadist)

Create a new distribution by applying the function post_processing to each atom of probadist (the weights remain the same).

InferOpt.compress_distribution! — Method

compress_distribution!(probadist[; atol])

Remove duplicated atoms in probadist (up to a tolerance on equality).

InferOpt.compute_expectation — Function

compute_expectation(probadist[, post_processing=identity])

Compute the expectation of post_processing(X) where X is a random variable distributed according to probadist.

This operation is made differentiable thanks to a custom reverse rule, even when post_processing itself is not a differentiable function.

Warning

Derivatives are computed with respect to probadist.weights only, assuming that probadist.atoms doesn't change (hence the name FixedAtomsProbabilityDistribution).

InferOpt.compute_probability_distribution — Function

compute_probability_distribution(layer, θ)

Apply a probabilistic layer (regularized or perturbed) to an objective direction θ in order to generate a FixedAtomsProbabilityDistribution on the vertices of a polytope.

The following layer types are supported:

AbstractPerturbed

RegularizedGeneric

InferOpt.Pushforward — Type

Pushforward{L,G}

Differentiable pushforward of a probabilistic layer with an arbitrary function post_processing.

Pushforward can be used for direct regret minimization (aka learning by experience) when the post-processing returns a cost.

Fields

layer::L: anything that implements compute_probability_distribution(layer, θ; kwargs...)

post_processing::P: callable

See also: FixedAtomsProbabilityDistribution.

InferOpt.Pushforward — Method

(pushforward::Pushforward)(θ)

Output the expectation of pushforward.post_processing(X), where X follows the distribution defined by pushforward.layer applied to θ.

Unlike compute_probability_distribution(pushforward, θ), this function is differentiable, even if pushforward.post_processing isn't.

See also: compute_expectation.

InferOpt.compute_probability_distribution — Method

compute_probability_distribution(pushforward, θ)

Output the distribution of pushforward.post_processing(X), where X follows the distribution defined by pushforward.layer applied to θ.

This function is not differentiable if pushforward.post_processing isn't.

See also: apply_on_atoms.

Interpolation

Reference

Differentiation of Blackbox Combinatorial Solvers

InferOpt.Interpolation — Type

Interpolation{F}

Piecewise-linear interpolation of a black-box optimizer.

Fields

maximizer::F: underlying argmax function

λ::Float64: smoothing parameter (smaller = more faithful approximation, larger = more informative gradients)

Reference: https://arxiv.org/abs/1912.02175

Smart "Predict, then Optimize"

Reference

Smart "Predict, then Optimize"

InferOpt.SPOPlusLoss — Type

SPOPlusLoss{F}

Convex surrogate of the Smart "Predict-then-Optimize" loss.

Fields

maximizer::F: linear maximizer function of the form θ ⟼ ŷ(θ) = argmax θᵀy

α::Float64: convexification parameter

Reference: https://arxiv.org/abs/1710.08005

Structured Support Vector Machines

Reference

Structured learning and prediction in computer vision, Chapter 6

InferOpt.IsBaseLoss — Type

IsBaseLoss{L}

Trait-based interface for loss functions δ(y, y_true), which are the base of the more complex StructuredSVMLoss.

For δ::L to comply with this interface, the following methods must exist:

(δ)(y, y_true)

compute_maximizer(δ, θ, α, y_true)

Available implementations

ZeroOneBaseLoss

InferOpt.compute_maximizer — Method

compute_maximizer(δ, θ, α, y_true)

Compute argmax_y {δ(y, y_true) + α θᵀ(y - y_true)} to deduce the gradient of a StructuredSVMLoss.

InferOpt.StructuredSVMLoss — Type

StructuredSVMLoss{L}

Loss associated with the Structured Support Vector Machine.

SSVM(θ, y_true) = max_y {base_loss(y, y_true) + α θᵀ(y - y_true)}

Fields

base_loss::L: of the IsBaseLoss trait

α::Float64

Reference: http://www.nowozin.net/sebastian/papers/nowozin2011structured-tutorial.pdf (Chapter 6)

InferOpt.ZeroOneBaseLoss — Type

ZeroOneBaseLoss

0-1 loss for multiclass classification: δ(y, y_true) = 0 if y = y_true, and 1 otherwise.

Frank-Wolfe

Reference

Efficient and Modular Implicit Differentiation

Reference

FrankWolfe.jl: a high-performance and flexible toolbox for Frank-Wolfe algorithms and Conditional Gradients

InferOpt.DEFAULT_FRANK_WOLFE_KWARGS — Constant

DEFAULT_FRANK_WOLFE_KWARGS

Default configuration for the Frank-Wolfe wrapper.

Parameters

away_steps=true: activate away steps to avoid zig-zagging

epsilon=1e-4: precision

lazy=true: caching strategy

line_search=FrankWolfe.Agnostic(): step size selection

max_iteration=10: number of iterations

timeout=1.0: maximum time in seconds

verbose=false: console output

InferOpt.LMOWrapper — Type

LMOWrapper{F,K}

Wraps a linear maximizer as a FrankWolfe.LinearMinimizationOracle.

Fields

maximizer::F: black box linear maximizer

maximizer_kwargs::K: keyword arguments passed to the maximizer whenever it is called

FrankWolfe.compute_extreme_point — Method

FrankWolfe.compute_extreme_point(lmo_wrapper::LMOWrapper, direction)

InferOpt.DifferentiableFrankWolfe — Type

DifferentiableFrankWolfe{F,G,M,S}

Parameterized version of the Frank-Wolfe algorithm θ -> argmin_{x ∈ C} f(x, θ), which can be differentiated implicitly wrt θ.

Fields

f::F: function f(x, θ) to minimize wrt x

f_grad1::G: gradient ∇ₓf(x, θ) of f wrt x

lmo::M: linear minimization oracle θ -> argmin_{x ∈ C} θᵀx, implicitly defines the polytope C

linear_solver::S: solver for linear systems of equations, used during implicit differentiation

Applicable methods

compute_probability_distribution(dfw::DifferentiableFrankWolfe, θ, x0)

(dfw::DifferentiableFrankWolfe)(θ, x0)

InferOpt.DifferentiableFrankWolfe — Method

(dfw::DifferentiableFrankWolfe)(θ, x0[; fw_kwargs=(;)])

Apply compute_probability_distribution(dfw, θ, x0) and return the expectation.

InferOpt.compute_probability_distribution — Method

compute_probability_distribution(dfw::DifferentiableFrankWolfe, θ, x0[; fw_kwargs=(;)])

Compute the optimal active set by applying the away-step Frank-Wolfe algorithm with initial point x0, then turn it into a probability distribution.

The named tuple fw_kwargs is passed as keyword arguments to FrankWolfe.away_frank_wolfe.

Regularized optimizers

Reference

Learning with Fenchel-Young Losses

InferOpt.IsRegularized — Type

IsRegularized{P}

Trait-based interface for regularized prediction functions ŷ(θ) = argmax {θᵀy - Ω(y)}.

For predictor::P to comply with this interface, the following methods must exist:

(predictor)(θ)

compute_regularization(predictor, y)

Available implementations

one_hot_argmax

soft_argmax

sparse_argmax

RegularizedGeneric

InferOpt.compute_regularization — Function

compute_regularization(predictor::P, y)

Compute the convex regularization function Ω(y).

InferOpt.RegularizedGeneric — Type

RegularizedGeneric{M,RF,RG,F,G,S}

Differentiable regularized prediction function ŷ(θ) = argmax_{y ∈ C} {θᵀy - Ω(y)}.

Relies on the Frank-Wolfe algorithm to minimize a concave objective on a polytope.

Fields

maximizer::M: linear maximization oracle θ -> argmax_{x ∈ C} θᵀx, implicitly defines the polytope C

Ω::RF: regularization function Ω(y)

Ω_grad::RG: gradient of the regularization function ∇Ω(y)

f::F: objective function f(x, θ) = Ω(y) - θᵀy minimized by Frank-Wolfe (computed automatically)

f_grad1::G: gradient of the objective function ∇ₓf(x, θ) = ∇Ω(y) - θ with respect to x (computed automatically)

linear_solver::S: solver for linear systems of equations, used during implicit differentiation

Applicable methods

compute_probability_distribution(regularized::RegularizedGeneric, θ)

(regularized::RegularizedGeneric)(θ)

See also: DifferentiableFrankWolfe.

InferOpt.RegularizedGeneric — Method

(regularized::RegularizedGeneric)(θ[; maximizer_kwargs=(;), fw_kwargs=(;)])

Apply compute_probability_distribution(regularized, θ) and return the expectation.

InferOpt.RegularizedGeneric — Method

RegularizedGeneric(maximizer[; Ω, Ω_grad, linear_solver=gmres])

Shorter constructor with defaults.

InferOpt.compute_probability_distribution — Method

compute_probability_distribution(regularized::RegularizedGeneric, θ[; maximizer_kwargs=(;), fw_kwargs=(;)])

Construct a DifferentiableFrankWolfe struct and call compute_probability_distribution on it.

The named tuple maximizer_kwargs is passed as keyword arguments to the underlying maximizer, which is wrapped inside a LMOWrapper. The named tuple fw_kwargs is passed as keyword arguments to FrankWolfe.away_frank_wolfe.

InferOpt.half_square_norm — Method

half_square_norm(x)

Compute the squared Euclidean norm of x and divide it by 2.

InferOpt.isproba — Method

isproba(x)

Check whether x ∈ [0,1].

InferOpt.isprobadist — Method

isprobadist(p)

Check whether the elements of p are nonnegative and sum to 1.

InferOpt.one_hot_argmax — Method

one_hot_argmax(z)

One-hot encoding of the argmax function.

InferOpt.positive_part — Method

positive_part(x)

Compute max(x,0).

InferOpt.ranking — Method

ranking(θ[; rev])

Compute the vector r such that rᵢ is the rank of θᵢ in θ.

InferOpt.shannon_entropy — Method

shannon_entropy(p)

Compute the Shannon entropy of a probability distribution: H(p) = -∑ pᵢlog(pᵢ).

InferOpt.soft_argmax — Method

soft_argmax(z)

Soft argmax activation function s(z) = (e^zᵢ / ∑ e^zⱼ)ᵢ.

Corresponds to regularized prediction on the probability simplex with entropic penalty.

InferOpt.simplex_projection_and_support — Method

simplex_projection_and_support(z)

Compute the Euclidean projection p of z on the probability simplex (also called sparse_argmax), and the indicators s of its support.

Reference: https://arxiv.org/abs/1602.02068.

InferOpt.sparse_argmax — Method

sparse_argmax(z)

Compute the Euclidean projection of the vector z onto the probability simplex.

Corresponds to regularized prediction on the probability simplex with square norm penalty.

Perturbed optimizers

Reference

Learning with Differentiable Perturbed Optimizers

InferOpt.AbstractPerturbed — Type

AbstractPerturbed{B}

Differentiable perturbation of a black box optimizer. The parameter parallel is a boolean value, equal to true if the perturbations are run in parallel.

Applicable functions

compute_probability_distribution(perturbed::AbstractPerturbed, θ)

(perturbed::AbstractPerturbed)(θ)

Available subtypes

PerturbedAdditive

PerturbedMultiplicative

These subtypes share the following fields:

maximizer: black box optimizer

ε: magnitude of the perturbation

nb_samples::Int: number of random samples for Monte-Carlo computations

rng::AbstractRNG: random number generator

seed::Union{Nothing,Int}: random seed

InferOpt.AbstractPerturbed — Method

(perturbed::AbstractPerturbed)(θ)

Apply compute_probability_distribution(perturbed, θ) and return the expectation.

InferOpt.compute_probability_distribution — Method

compute_probability_distribution(perturbed::AbstractPerturbed, θ)

Turn random perturbations of θ into a distribution on polytope vertices.

InferOpt.sample_perturbations — Method

sample_perturbations(perturbed::AbstractPerturbed, θ)

Draw random perturbations Z which will be applied to the objective direction θ.

InferOpt.PerturbedAdditive — Type

PerturbedAdditive{F}

Differentiable normal perturbation of a black-box optimizer of type F: the input undergoes θ -> θ + εZ where Z ∼ N(0, I).

See also: AbstractPerturbed.

Reference: https://arxiv.org/abs/2002.08676

InferOpt.PerturbedAdditive — Method

PerturbedAdditive(maximizer[; ε=1.0, nb_samples=1])

Shorter constructor with defaults.

InferOpt.PerturbedMultiplicative — Type

PerturbedMultiplicative{F}

Differentiable log-normal perturbation of a black-box optimizer of type F: the input undergoes θ -> θ ⊙ exp[εZ - ε²/2] where Z ∼ N(0, I).

See also: AbstractPerturbed.

Reference: preprint coming soon.

InferOpt.PerturbedMultiplicative — Method

PerturbedMultiplicative(maximizer[; ε=1.0, nb_samples=1])

Shorter constructor with defaults.

Fenchel-Young losses

Reference

Learning with Fenchel-Young Losses

InferOpt.FenchelYoungLoss — Type

FenchelYoungLoss{P}

Fenchel-Young loss associated with a given regularized prediction function.

Fields

predictor::P: prediction function of the form ŷ(θ) = argmax {θᵀy - Ω(y)}

Reference: https://arxiv.org/abs/1901.02324

Generalized imitation losses

Reference

Learning with Combinatorial Optimization Layers: a Probabilistic Approach

InferOpt.ImitationLoss — Type

ImitationLoss{L,R,P}

Generic imitation loss: maxy baseloss(y, ttrue) + α θᵀ(y - ytrue) - (Ω(y) - Ω(y_true))).

When base_loss = 0, this loss is equivalent to a FenchelYoungLoss. When Ω = 0, this loss is equivalent to the StructuredSVMLoss.

Fields

maximizer::P: function that computes argmaxy baseloss(y, ttrue) + α θᵀ(y - ytrue) - (Ω(y) - Ω(ytrue)), takes (θ, ytrue, kwargs...) or (θ, t_true, kwargs...) as input

base_loss::L: base loss, takes (y, t_true) as input

Ω::R: regularization, takes y as input

α::Float64: default value of 1.0

Note: by default, t_true is a named tuple with field y_true, but can be any data structure for which the get_y_true method is implemented.

InferOpt.ImitationLoss — Method

ImitationLoss(maximizer[; base_loss=(y,t_true)->0.0, Ω=y->0.0, α=1.0])

Shorter constructor with defaults.

InferOpt.get_y_true — Method

get_y_true(t_true::Any)

Retrieve y_true from t_true. This method should be implemented when using a custom data structure for t_true other than a NamedTuple.

InferOpt.get_y_true — Method

get_y_true(t_true::NamedTuple)

Retrieve y_true from t_true. t_true must contain an y_true field.

Index

InferOpt.DEFAULT_FRANK_WOLFE_KWARGS

InferOpt.AbstractPerturbed

InferOpt.AbstractPerturbed

InferOpt.DifferentiableFrankWolfe

InferOpt.DifferentiableFrankWolfe

InferOpt.FenchelYoungLoss

InferOpt.FixedAtomsProbabilityDistribution

InferOpt.FixedAtomsProbabilityDistribution

InferOpt.ImitationLoss

InferOpt.ImitationLoss

InferOpt.Interpolation

InferOpt.IsBaseLoss

InferOpt.IsRegularized

InferOpt.LMOWrapper

InferOpt.PerturbedAdditive

InferOpt.PerturbedAdditive

InferOpt.PerturbedMultiplicative

InferOpt.PerturbedMultiplicative

InferOpt.Pushforward

InferOpt.Pushforward

InferOpt.RegularizedGeneric

InferOpt.RegularizedGeneric

InferOpt.RegularizedGeneric

InferOpt.SPOPlusLoss

InferOpt.StructuredSVMLoss

InferOpt.ZeroOneBaseLoss

InferOpt.apply_on_atoms

InferOpt.compress_distribution!

InferOpt.compute_expectation

InferOpt.compute_maximizer

InferOpt.compute_probability_distribution

InferOpt.compute_probability_distribution

InferOpt.compute_probability_distribution

InferOpt.compute_probability_distribution

InferOpt.compute_probability_distribution

InferOpt.compute_regularization

InferOpt.get_y_true

InferOpt.get_y_true

InferOpt.half_square_norm

InferOpt.isproba

InferOpt.isprobadist

InferOpt.one_hot_argmax

InferOpt.positive_part

InferOpt.ranking

InferOpt.sample_perturbations

InferOpt.shannon_entropy

InferOpt.simplex_projection_and_support

InferOpt.soft_argmax

InferOpt.sparse_argmax