API Reference

InferOpt — Module

InferOpt

A toolbox for using combinatorial optimization algorithms within machine learning pipelines.

See our preprint https://arxiv.org/abs/2207.13513

Types

InferOpt.AbstractLayer — Type

AbstractLayer

Supertype for all the layers defined in InferOpt.

All of these layers are callable, and differentiable with any ChainRules-compatible autodiff backend.

Interface

(layer::AbstractLayer)(args...; kwargs...)

InferOpt.AbstractLossLayer — Type

AbstractLossLayer <: AbstractLayer

Supertype for all the loss layers defined in InferOpt.

Depending on the precise loss, the arguments to the layer might vary

Interface

(layer::AbstractLossLayer)(θ; kwargs...) or

(layer::AbstractLossLayer)(θ, θ_true; kwargs...) or

(layer::AbstractLossLayer)(θ, y_true; kwargs...) or

(layer::AbstractLossLayer)(θ, (; θ_true, y_true); kwargs...)

InferOpt.AbstractOptimizationLayer — Type

AbstractOptimizationLayer <: AbstractLayer

Supertype for all the optimization layers defined in InferOpt.

Interface

(layer::AbstractOptimizationLayer)(θ; kwargs...)

compute_probability_distribution(layer, θ; kwargs...) (only if the layer is probabilistic)

InferOpt.AbstractPerturbed — Type

AbstractPerturbed{parallel} <: AbstractOptimizationLayer

Differentiable perturbation of a black box optimizer.

The parameter parallel is a boolean value, equal to true if the perturbations are run in parallel.

Available implementations

PerturbedAdditive

PerturbedMultiplicative

These two subtypes share the following fields:

maximizer: black box optimizer

ε: magnitude of the perturbation

nb_samples::Int: number of random samples for Monte-Carlo computations

rng::AbstractRNG: random number generator

seed::Union{Nothing,Int}: random seed

InferOpt.AbstractPerturbed — Method

(perturbed::AbstractPerturbed)(θ; kwargs...)

Apply compute_probability_distribution(perturbed, θ; kwargs...) and return the expectation.

InferOpt.AbstractRegularized — Type

AbstractRegularized{parallel} <: AbstractOptimizationLayer

Convex regularization perturbation of a black box optimizer

ŷ(θ) = argmax_{y ∈ C} {θᵀy - Ω(y)}

Interface

(regularized::AbstractRegularized)(θ; kwargs...): return ŷ(θ)

compute_regularization(regularized, y): return Ω(y)

Available implementations

SoftArgmax

SparseArgmax

RegularizedFrankWolfe

InferOpt.FenchelYoungLoss — Type

FenchelYoungLoss <: AbstractLossLayer

Fenchel-Young loss associated with a given optimization layer.

L(θ, y_true) = (Ω(y_true) - θᵀy_true) - (Ω(ŷ) - θᵀŷ)

Reference: https://arxiv.org/abs/1901.02324

Fields

optimization_layer::AbstractOptimizationLayer: optimization layer that can be formulated as ŷ(θ) = argmax {θᵀy - Ω(y)} (either regularized or perturbed)

InferOpt.FenchelYoungLoss — Method

(fyl::FenchelYoungLoss)(θ, y_true; kwargs...)

InferOpt.FixedAtomsProbabilityDistribution — Type

FixedAtomsProbabilityDistribution{A,W}

Encodes a probability distribution with finite support and fixed atoms.

See compute_expectation to understand the name of this struct.

Fields

atoms::Vector{A}: elements of the support

weights::Vector{W}: probability values for each atom (must sum to 1)

InferOpt.IdentityRelaxation — Type

IdentityRelaxation <: AbstractOptimizationLayer

Naive relaxation of a black-box optimizer where constraints are simply forgotten.

Consider (centering and) normalizing θ before applying it.

Fields

maximizer: underlying argmax function

Reference: https://arxiv.org/abs/2205.15213

InferOpt.ImitationLoss — Type

ImitationLoss <: AbstractLossLayer

Generic imitation loss of the form

L(θ, t_true) = max_y {δ(y, t_true) + α θᵀ(y - y_true) - (Ω(y) - Ω(y_true))}

When δ is zero, this is equivalent to a FenchelYoungLoss.

When Ω is zero, this is equivalent to a StructuredSVMLoss.

Note: by default, t_true is a named tuple with field y_true, but it can be any data structure for which the get_y_true method is implemented.

Fields

aux_loss_maximizer: function of (θ, t_true, α) that computes the argmax in the problem above

δ: base loss function

Ω: regularization function

α::Float64: hyperparameter with a default value of 1.0

InferOpt.ImitationLoss — Method

(il::ImitationLoss)(θ, t_true; kwargs...)

InferOpt.ImitationLoss — Method

ImitationLoss(; aux_loss_maximizer, δ, Ω, α=1.0)

Explicit constructor with keyword arguments.

InferOpt.Interpolation — Type

Interpolation <: AbstractOptimizationLayer

Piecewise-linear interpolation of a black-box optimizer.

Fields

maximizer: underlying argmax function

λ::Float64: smoothing parameter (smaller = more faithful approximation, larger = more informative gradients)

Reference: https://arxiv.org/abs/1912.02175

InferOpt.PerturbedAdditive — Type

PerturbedAdditive <: AbstractPerturbed

Differentiable normal perturbation of a black-box maximizer: the input undergoes θ -> θ + εZ where Z ∼ N(0, I).

Reference: https://arxiv.org/abs/2002.08676

See AbstractPerturbed for more details.

InferOpt.PerturbedAdditive — Method

PerturbedAdditive(maximizer[; ε=1.0, nb_samples=1])

InferOpt.PerturbedMultiplicative — Type

PerturbedMultiplicative <: AbstractPerturbed

Differentiable log-normal perturbation of a black-box maximizer: the input undergoes θ -> θ ⊙ exp[εZ - ε²/2] where Z ∼ N(0, I).

Reference: https://arxiv.org/abs/2207.13513

See AbstractPerturbed for more details.

InferOpt.PerturbedMultiplicative — Method

PerturbedMultiplicative(maximizer[; ε=1.0, nb_samples=1])

InferOpt.Pushforward — Type

Pushforward <: AbstractLayer

Differentiable pushforward of a probabilistic optimization layer with an arbitrary function post-processing function.

Pushforward can be used for direct regret minimization (aka learning by experience) when the post-processing returns a cost.

Fields

optimization_layer::AbstractOptimizationLayer: probabilistic optimization layer

post_processing: callable

See also: FixedAtomsProbabilityDistribution.

InferOpt.Pushforward — Method

(pushforward::Pushforward)(θ; kwargs...)

Output the expectation of pushforward.post_processing(X), where X follows the distribution defined by pushforward.optimization_layer applied to θ.

Unlike compute_probability_distribution(pushforward, θ), this function is differentiable, even if pushforward.post_processing isn't.

See also: compute_expectation.

InferOpt.RegularizedFrankWolfe — Type

RegularizedFrankWolfe <: AbstractRegularized

Regularized optimization layer which relies on the Frank-Wolfe algorithm to define a probability distribution while solving

ŷ(θ) = argmax_{y ∈ C} {θᵀy - Ω(y)}

Warning

Since this is a conditional dependency, you need to have loaded the package DifferentiableFrankWolfe.jl before using RegularizedFrankWolfe.

Fields

linear_maximizer: linear maximization oracle θ -> argmax_{x ∈ C} θᵀx, implicitly defines the polytope C

Ω: regularization function Ω(y)

Ω_grad: gradient function of the regularization function ∇Ω(y)

frank_wolfe_kwargs: named tuple of keyword arguments passed to the Frank-Wolfe algorithm

Frank-Wolfe parameters

Some values you can tune:

epsilon::Float64: precision target

max_iteration::Integer: max number of iterations

timeout::Float64: max runtime in seconds

lazy::Bool: caching strategy

away_steps::Bool: avoid zig-zagging

line_search::FrankWolfe.LineSearchMethod: step size selection

verbose::Bool: console output

See the documentation of FrankWolfe.jl for details.

InferOpt.RegularizedFrankWolfe — Method

(regularized::RegularizedFrankWolfe)(θ; kwargs...)

Apply compute_probability_distribution(regularized, θ; kwargs...) and return the expectation.

InferOpt.RegularizedFrankWolfe — Method

RegularizedFrankWolfe(linear_maximizer; Ω, Ω_grad, frank_wolfe_kwargs=(;))

InferOpt.SPOPlusLoss — Type

SPOPlusLoss <: AbstractLossLayer

Convex surrogate of the Smart "Predict-then-Optimize" loss.

Fields

maximizer: linear maximizer function of the form θ -> ŷ(θ) = argmax θᵀy

α::Float64: convexification parameter, default = 2.0

Reference: https://arxiv.org/abs/1710.08005

InferOpt.SPOPlusLoss — Method

(spol::SPOPlusLoss)(θ, θ_true, y_true; kwargs...)

InferOpt.SPOPlusLoss — Method

(spol::SPOPlusLoss)(θ, θ_true; kwargs...)

InferOpt.SPOPlusLoss — Method

SPOPlusLoss(maximizer; α=2.0)

InferOpt.SoftArgmax — Type

SoftArgmax <: Regularized

Soft argmax activation function s(z) = (e^zᵢ / ∑ e^zⱼ)ᵢ.

Corresponds to regularized prediction on the probability simplex with entropic penalty.

InferOpt.SparseArgmax — Type

SparseArgmax <: AbstractRegularized

Compute the Euclidean projection of the vector z onto the probability simplex.

Corresponds to regularized prediction on the probability simplex with square norm penalty.

InferOpt.StructuredSVMLoss — Type

StructuredSVMLoss <: AbstractLossLayer

Loss associated with the Structured Support Vector Machine, defined by

L(θ, y_true) = max_y {δ(y, y_true) + α θᵀ(y - y_true)}

Reference: http://www.nowozin.net/sebastian/papers/nowozin2011structured-tutorial.pdf (Chapter 6)

Fields

aux_loss_maximizer::M: function of (θ, y_true, α) that computes the argmax in the problem above

δ::L: base loss function

α::Float64: hyperparameter with a default value of 1.0

InferOpt.StructuredSVMLoss — Method

(ssvml::StructuredSVMLoss)(θ, y_true; kwargs...)

InferOpt.StructuredSVMLoss — Method

StructuredSVMLoss(; aux_loss_maximizer, δ, α=1.0)

Functions

Base.rand — Method

rand([rng,] probadist)

Sample from the atoms of probadist according to their weights.

InferOpt.ZeroOneImitationLoss — Function

ZeroOneStructuredSVMLoss(α)

Implementation of the ImitationLoss based on a 0-1 loss for multiclass classification with no regularization.

InferOpt.ZeroOneStructuredSVMLoss — Function

ZeroOneStructuredSVMLoss

Implementation of the StructuredSVMLoss based on a 0-1 loss for multiclass classification.

InferOpt.apply_on_atoms — Method

apply_on_atoms(post_processing, probadist)

Create a new distribution by applying the function post_processing to each atom of probadist (the weights remain the same).

InferOpt.compute_expectation — Function

compute_expectation(probadist[, post_processing=identity])

Compute the expectation of post_processing(X) where X is a random variable distributed according to probadist.

This operation is made differentiable thanks to a custom reverse rule, even when post_processing itself is not a differentiable function.

Warning

Derivatives are computed with respect to probadist.weights only, assuming that probadist.atoms doesn't change (hence the name FixedAtomsProbabilityDistribution).

InferOpt.compute_probability_distribution — Function

compute_probability_distribution(layer, θ; kwargs...)

Apply a probabilistic optimization layer to an objective direction θ in order to generate a FixedAtomsProbabilityDistribution on the vertices of a polytope.

InferOpt.compute_probability_distribution — Method

compute_probability_distribution(perturbed::AbstractPerturbed, θ; kwargs...)

Turn random perturbations of θ into a distribution on polytope vertices.

Keyword arguments are passed to the underlying linear maximizer.

InferOpt.compute_probability_distribution — Method

compute_probability_distribution(pushforward, θ)

Output the distribution of pushforward.post_processing(X), where X follows the distribution defined by pushforward.optimization_layer applied to θ.

This function is not differentiable if pushforward.post_processing isn't.

See also: apply_on_atoms.

InferOpt.compute_regularization — Function

compute_regularization(regularized, y)

Return the convex penalty Ω(y) associated with an AbstractRegularized layer.

InferOpt.get_y_true — Function

get_y_true(t_true::Any)

Retrieve y_true from t_true.

This method should be implemented when using a custom data structure for t_true other than a NamedTuple.

InferOpt.get_y_true — Method

get_y_true(t_true::NamedTuple)

Retrieve y_true from t_true. t_true must contain an y_true field.

InferOpt.half_square_norm — Method

half_square_norm(x)

Compute the squared Euclidean norm of x and divide it by 2.

InferOpt.isproba — Method

isproba(x)

Check whether x ∈ [0,1].

InferOpt.isprobadist — Method

isprobadist(p)

Check whether the elements of p are nonnegative and sum to 1.

InferOpt.one_hot_argmax — Method

one_hot_argmax(z)

One-hot encoding of the argmax function.

InferOpt.positive_part — Method

positive_part(x)

Compute max(x,0).

InferOpt.ranking — Method

ranking(θ[; rev])

Compute the vector r such that rᵢ is the rank of θᵢ in θ.

InferOpt.sample_perturbations — Method

sample_perturbations(perturbed::AbstractPerturbed, θ)

Draw random perturbations Z which will be applied to the objective direction θ.

InferOpt.shannon_entropy — Method

shannon_entropy(p)

Compute the Shannon entropy of a probability distribution: H(p) = -∑ pᵢlog(pᵢ).

InferOpt.simplex_projection_and_support — Method

simplex_projection_and_support(z)

Compute the Euclidean projection p of z on the probability simplex (also called sparse_argmax), and the indicators s of its support.

Reference: https://arxiv.org/abs/1602.02068.

InferOpt.zero_one_loss — Method

zero_one_loss(y, y_true)

0-1 loss for multiclass classification: δ(y, y_true) = 0 if y = y_true, and 1 otherwise.

InferOpt.zero_one_loss_maximizer — Method

zero_one_loss_maximizer(y, y_true; α)

For δ = zero_one_loss, compute

argmax_y {δ(y, y_true) + α θᵀ(y - y_true)}

Index

InferOpt.AbstractLayer

InferOpt.AbstractLossLayer

InferOpt.AbstractOptimizationLayer

InferOpt.AbstractPerturbed

InferOpt.AbstractPerturbed

InferOpt.AbstractRegularized

InferOpt.FenchelYoungLoss

InferOpt.FenchelYoungLoss

InferOpt.FixedAtomsProbabilityDistribution

InferOpt.IdentityRelaxation

InferOpt.ImitationLoss

InferOpt.ImitationLoss

InferOpt.ImitationLoss

InferOpt.Interpolation

InferOpt.PerturbedAdditive

InferOpt.PerturbedAdditive

InferOpt.PerturbedMultiplicative

InferOpt.PerturbedMultiplicative

InferOpt.Pushforward

InferOpt.Pushforward

InferOpt.RegularizedFrankWolfe

InferOpt.RegularizedFrankWolfe

InferOpt.RegularizedFrankWolfe

InferOpt.SPOPlusLoss

InferOpt.SPOPlusLoss

InferOpt.SPOPlusLoss

InferOpt.SPOPlusLoss

InferOpt.SoftArgmax

InferOpt.SparseArgmax

InferOpt.StructuredSVMLoss

InferOpt.StructuredSVMLoss

InferOpt.StructuredSVMLoss

InferOpt.ZeroOneImitationLoss

InferOpt.ZeroOneStructuredSVMLoss

InferOpt.apply_on_atoms

InferOpt.compute_expectation

InferOpt.compute_probability_distribution

InferOpt.compute_probability_distribution

InferOpt.compute_probability_distribution

InferOpt.compute_regularization

InferOpt.get_y_true

InferOpt.get_y_true

InferOpt.half_square_norm

InferOpt.isproba

InferOpt.isprobadist

InferOpt.one_hot_argmax

InferOpt.positive_part

InferOpt.ranking

InferOpt.sample_perturbations

InferOpt.shannon_entropy

InferOpt.simplex_projection_and_support

InferOpt.zero_one_loss

InferOpt.zero_one_loss_maximizer